requirements.txt
# Core
torch>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Encoders and LMs
transformers>=4.36.0
sentence-transformers>=2.2.0

# Data
datasets>=2.16.0    # HuggingFace datasets for WikiText-103

# Nearest neighbors
faiss-gpu>=1.7.4    # Use faiss-cpu if no GPU

# Config and logging
pyyaml>=6.0
tqdm>=4.66.0
wandb>=0.16.0       # Optional: experiment tracking

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
umap-learn>=0.5.0   # For latent space visualizations

# Utilities
scikit-learn>=1.3.0
